+++ appkit/animatable_property_container.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSAnimatablePropertyContainerAdaptor : NSObject <NSAnimatablePropertyContainer>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapAnimatablePropertyContainer(long goID);
+void* WrapAnimatablePropertyContainer(uintptr_t goID);
