+++ appkit/animation_delegate.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSAnimationDelegateAdaptor : NSObject <NSAnimationDelegate>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapAnimationDelegate(long goID);
+void* WrapAnimationDelegate(uintptr_t goID);
