+++ appkit/application_delegate.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSApplicationDelegateAdaptor : NSObject <NSApplicationDelegate>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapApplicationDelegate(long goID);
+void* WrapApplicationDelegate(uintptr_t goID);
