+++ appkit/collection_view_delegate.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSCollectionViewDelegateAdaptor : NSObject <NSCollectionViewDelegate>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapCollectionViewDelegate(long goID);
+void* WrapCollectionViewDelegate(uintptr_t goID);
