+++ appkit/collection_view_delegate_flow_layout.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSCollectionViewDelegateFlowLayoutAdaptor : NSObject <NSCollectionViewDelegateFlowLayout>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapCollectionViewDelegateFlowLayout(long goID);
+void* WrapCollectionViewDelegateFlowLayout(uintptr_t goID);
