+++ appkit/combo_box_data_source.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSComboBoxDataSourceAdaptor : NSObject <NSComboBoxDataSource>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapComboBoxDataSource(long goID);
+void* WrapComboBoxDataSource(uintptr_t goID);
