+++ appkit/combo_box_delegate.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSComboBoxDelegateAdaptor : NSObject <NSComboBoxDelegate>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapComboBoxDelegate(long goID);
+void* WrapComboBoxDelegate(uintptr_t goID);
