+++ appkit/dragging_destination.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSDraggingDestinationAdaptor : NSObject <NSDraggingDestination>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapDraggingDestination(long goID);
+void* WrapDraggingDestination(uintptr_t goID);
