+++ appkit/dragging_info.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSDraggingInfoAdaptor : NSObject <NSDraggingInfo>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapDraggingInfo(long goID);
+void* WrapDraggingInfo(uintptr_t goID);
