+++ appkit/dragging_source.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSDraggingSourceAdaptor : NSObject <NSDraggingSource>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapDraggingSource(long goID);
+void* WrapDraggingSource(uintptr_t goID);
