+++ appkit/font_changing.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSFontChangingAdaptor : NSObject <NSFontChanging>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapFontChanging(long goID);
+void* WrapFontChanging(uintptr_t goID);
