+++ appkit/outline_view_delegate.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSOutlineViewDelegateAdaptor : NSObject <NSOutlineViewDelegate>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapOutlineViewDelegate(long goID);
+void* WrapOutlineViewDelegate(uintptr_t goID);
