+++ appkit/path_cell_delegate.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSPathCellDelegateAdaptor : NSObject <NSPathCellDelegate>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapPathCellDelegate(long goID);
+void* WrapPathCellDelegate(uintptr_t goID);
