+++ appkit/path_control_delegate.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSPathControlDelegateAdaptor : NSObject <NSPathControlDelegate>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapPathControlDelegate(long goID);
+void* WrapPathControlDelegate(uintptr_t goID);
