+++ appkit/spring_loading_destination.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSSpringLoadingDestinationAdaptor : NSObject <NSSpringLoadingDestination>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapSpringLoadingDestination(long goID);
+void* WrapSpringLoadingDestination(uintptr_t goID);
