+++ appkit/tab_view_delegate.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSTabViewDelegateAdaptor : NSObject <NSTabViewDelegate>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapTabViewDelegate(long goID);
+void* WrapTabViewDelegate(uintptr_t goID);
