+++ appkit/table_view_delegate.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSTableViewDelegateAdaptor : NSObject <NSTableViewDelegate>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapTableViewDelegate(long goID);
+void* WrapTableViewDelegate(uintptr_t goID);
