+++ appkit/text_delegate.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSTextDelegateAdaptor : NSObject <NSTextDelegate>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapTextDelegate(long goID);
+void* WrapTextDelegate(uintptr_t goID);
