+++ appkit/text_field_delegate.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSTextFieldDelegateAdaptor : NSObject <NSTextFieldDelegate>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapTextFieldDelegate(long goID);
+void* WrapTextFieldDelegate(uintptr_t goID);
