+++ appkit/touch_bar_delegate.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSTouchBarDelegateAdaptor : NSObject <NSTouchBarDelegate>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapTouchBarDelegate(long goID);
+void* WrapTouchBarDelegate(uintptr_t goID);
