+++ appkit/window_delegate.h
@@ -1,10 +1,11 @@
 #import <stdbool.h>
+#import <stdint.h>
 #import <stdlib.h>
 #import <utils.h>
 #import <Appkit/Appkit.h>
 
 @interface NSWindowDelegateAdaptor : NSObject <NSWindowDelegate>
-@property (assign) long goID;
+@property (assign) uintptr_t goID;
 @end
 
-void* WrapWindowDelegate(long goID);
+void* WrapWindowDelegate(uintptr_t goID);
